chore: rearrange package structure
