+++ domain/src/main/kotlin/com/mlt/checkareercorespring/domain/samplenode/model/dto/RelationsSaveRequest.kt
@@ -1,0 +1,5 @@
+package com.mlt.checkareercorespring.domain.samplenode.model.dto
+
+data class RelationsSaveRequest(
+    val targetNodeIds: Set<Long>
+)
