+++ domain/src/main/kotlin/com/mlt/checkareercorespring/domain/samplenode/model/dto/SampleNodeSaveRequest.kt
@@ -1,0 +1,12 @@
+package com.mlt.checkareercorespring.domain.samplenode.model.dto
+
+import com.mlt.checkareercorespring.domain.samplenode.model.entity.SampleNode
+
+data class SampleNodeSaveRequest(
+    val id: Long,
+    val name: String
+) {
+    fun toSampleNode(): SampleNode {
+        return SampleNode(id, name)
+    }
+}
