+++ domain/src/main/kotlin/com/mlt/checkareercorespring/domain/user/model/dto/UserGetResponse.kt
@@ -1,0 +1,6 @@
+package com.mlt.checkareercorespring.domain.user.model.dto
+
+data class UserGetResponse(
+    val id: Long,
+    val name: String
+)
