+++ domain/src/main/kotlin/com/mlt/checkareercorespring/domain/user/model/dto/UserSaveRequest.kt
@@ -1,0 +1,7 @@
+package com.mlt.checkareercorespring.domain.user.model.dto
+
+data class UserSaveRequest(
+    val id: Long,
+    val name: String,
+    val skillIds: Set<Long>
+)
