+++ domain/src/main/kotlin/com/mlt/checkareercorespring/domain/user/model/dto/UserSkillUpdateRequest.kt
@@ -1,0 +1,5 @@
+package com.mlt.checkareercorespring.domain.user.model.dto
+
+data class UserSkillUpdateRequest(
+    val skillIds: Set<Long>
+)
