+++ web/src/main/kotlin/com/mlt/checkareercorespring/controller/HealthCheckController.kt
@@ -1,0 +1,14 @@
+package com.mlt.checkareercorespring.controller
+
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+class HealthCheckController {
+
+    @GetMapping("/health")
+    fun healthCheck(): ResponseEntity<String> {
+        return ResponseEntity.ok("I'm healthy")
+    }
+}
