--- web/src/main/kotlin/com/mlt/checkareercorespring/samplenode/domain/SampleNodeRepository.kt
@@ -1,5 +1,0 @@
-package com.mlt.checkareercorespring.samplenode.domain
-
-import org.springframework.data.neo4j.repository.Neo4jRepository
-
-interface SampleNodeRepository : Neo4jRepository<SampleNode, Long>
