--- web/src/main/kotlin/com/mlt/checkareercorespring/samplenode/dto/RelationsSaveRequest.kt
@@ -1,5 +1,0 @@
-package com.mlt.checkareercorespring.samplenode.dto
-
-data class RelationsSaveRequest(
-    val targetNodeIds: Set<Long>
-)
