--- web/src/main/kotlin/com/mlt/checkareercorespring/samplenode/dto/SampleNodeSaveRequest.kt
@@ -1,12 +1,0 @@
-package com.mlt.checkareercorespring.samplenode.dto
-
-import com.mlt.checkareercorespring.samplenode.domain.SampleNode
-
-data class SampleNodeSaveRequest(
-    val id: Long,
-    val name: String
-) {
-    fun toSampleNode(): SampleNode {
-        return SampleNode(id, name)
-    }
-}
