--- web/src/main/kotlin/com/mlt/checkareercorespring/skill/domain/Skill.kt
@@ -1,11 +1,0 @@
-package com.mlt.checkareercorespring.skill.domain
-
-import org.springframework.data.neo4j.core.schema.Id
-import org.springframework.data.neo4j.core.schema.Node
-
-@Node("Skill")
-class Skill(
-    @Id
-    val id: Long,
-    val name: String
-)
