--- web/src/main/kotlin/com/mlt/checkareercorespring/skill/domain/SkillRepository.kt
@@ -1,5 +1,0 @@
-package com.mlt.checkareercorespring.skill.domain
-
-import org.springframework.data.neo4j.repository.Neo4jRepository
-
-interface SkillRepository : Neo4jRepository<Skill, Long>
