--- web/src/main/kotlin/com/mlt/checkareercorespring/user/dto/UserGetResponse.kt
@@ -1,6 +1,0 @@
-package com.mlt.checkareercorespring.user.dto
-
-data class UserGetResponse(
-    val id: Long,
-    val name: String
-)
