--- web/src/main/kotlin/com/mlt/checkareercorespring/user/dto/UserSaveRequest.kt
@@ -1,7 +1,0 @@
-package com.mlt.checkareercorespring.user.dto
-
-data class UserSaveRequest(
-    val id: Long,
-    val name: String,
-    val skillIds: Set<Long>
-)
