--- web/src/main/kotlin/com/mlt/checkareercorespring/user/dto/UserSkillUpdateRequest.kt
@@ -1,5 +1,0 @@
-package com.mlt.checkareercorespring.user.dto
-
-data class UserSkillUpdateRequest(
-    val skillIds: Set<Long>
-)
